Split out type definitions
