+++ src/UserSerialized.ts
@@ -1,0 +1,1 @@
+export type UserSerialized = string
