+++ src/setAuthCookies.ts
@@ -3,8 +3,9 @@ import { setCookie } from 'src/cookies'
 import { getUserCookieName, getUserTokensCookieName } from 'src/authCookies'
 import { getConfig } from 'src/config'
 import logDebug from 'src/logDebug'
-import createUser, { User } from 'src/createUser'
+import createUser from 'src/createUser'
 import { NextApiRequest, NextApiResponse } from 'next'
+import { User } from './User'
 
 export type SetAuthCookies = (
   req: NextApiRequest,
